anomaly_detection/requirements.txt
flask==2.2.5
pandas==2.1.2
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.2
matplotlib==3.8.1
flask-cors==4.0.0
python-magic-bin==0.4.14
python-docx==0.8.11
PyPDF2==3.0.1
numpy>=1.21.0
openai>=1.0.0
python-dotenv>=0.19.0
